add draft and garbage state to task, along with all related functions

+++ src/main/java/com/alphacat/task/TaskServiceImpl.java
@@ -80,41 +80,70 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public int add(TaskVO taskVO) {
+        return add(taskVO, true);
+    }
+
+    @Override
+    public int add(TaskVO taskVO, boolean normal) {
         // set up new task id
         Integer tid = taskMapper.getNewId();
-        int id = tid == null ? 1 : tid;
+        final int id = tid == null ? 1 : tid;
         taskVO.setId(id);
         Task task = taskConverter.toPOJO(taskVO);
-        // check start time and end time
-        checkTaskDate(task);
-        // add task's basic info and settings
-        taskMapper.add(task);
-        // add labels
-        taskVO.getLabels().forEach(l -> labelMapper.add(taskConverter.toPOJO(l, id)));
-        // add a new task end job
-        scheduler.scheduleSingleJob(task);
+        if(normal) {
+            // check start time and end time
+            checkTaskDate(task);
+            // add task's basic info and settings
+            taskMapper.add(task);
+            taskMapper.setState(id, 1);
+            // add labels
+            taskVO.getLabels().forEach(l -> labelMapper.add(taskConverter.toPOJO(l, id)));
+            // add a new task end job
+            scheduler.scheduleSingleJob(task);
+        } else {
+            // save as draft
+            taskMapper.add(task);
+            taskMapper.setState(id, 0);
+            taskVO.getLabels().forEach(l -> labelMapper.add(taskConverter.toPOJO(l, id)));
+        }
         return id;
     }
 
     @Override
     public void update(TaskVO taskVO) {
-        int id = taskVO.getId();
+        update(taskVO, true);
+    }
+
+    @Override
+    public void update(TaskVO taskVO, boolean normal) {
+        final int id = taskVO.getId();
+        if(!normal) {
+            // updating draft doesn't need checking
+            taskMapper.update(taskConverter.toPOJO(taskVO));
+            taskMapper.setState(id, 0);
+            labelMapper.delete(id);
+            taskVO.getLabels().
+                    forEach(l -> labelMapper.add(taskConverter.toPOJO(l, id)));
+            return;
+        }
         Task origin = taskMapper.get(id);
+        int originalState = taskMapper.getState(id);
         Date now = new Date(Calendar.getInstance().getTimeInMillis());
-        // 1. change everything if it's not started
-        if(now.before(origin.getStartTime())) {
+        // case 1. change everything if it's not started or it was a draft
+        if(originalState == 0 || now.before(origin.getStartTime())) {
             Task task = taskConverter.toPOJO(taskVO);
             // check date first
             checkTaskDate(task);
             taskMapper.update(task);
-            labelMapper.delete(taskVO.getId());
+            taskMapper.setState(id, 1);
+            labelMapper.delete(id);
             taskVO.getLabels().
                     forEach(l -> labelMapper.add(taskConverter.toPOJO(l, id)));
             // then update the task end job
             scheduler.scheduleSingleJob(task);
             return;
         }
-        // 2. change only name, description and endTime if underway
+        // case 2. change only name, description and endTime if underway
         if(now.after(origin.getStartTime()) && now.before(origin.getEndTime())) {
             // check date first
             Date endTime = Date.valueOf(taskVO.getEndTime());
@@ -128,7 +157,7 @@ public class TaskServiceImpl implements TaskService {
             // update the task end job
             scheduler.scheduleSingleJob(origin);
         }
-        // 3. change nothing if ended
+        // case 3. change nothing if ended
     }
 
     /**
@@ -136,7 +165,59 @@ public class TaskServiceImpl implements TaskService {
      */
     @Override
     public void delete(int id) {
-        taskMapper.delete(id);
+        int state = taskMapper.getState(id);
+        if(state == 2) {
+            taskMapper.delete(id);
+        }
+    }
+
+    @Override
+    public void setToDraft(int id) {
+        int state = taskMapper.getState(id);
+        if(state == 0) {
+            return;
+        }
+        if(state == 2) {
+            taskMapper.setState(id, 0);
+            return;
+        }
+        Task task = taskMapper.get(id);
+        Date now = Date.valueOf(Calendar.getInstance().toString());
+        if(now.before(task.getStartTime())) {
+            taskMapper.setState(id, 0);
+        }
+    }
+
+    @Override
+    public void setToGarbage(int id) {
+        int state = taskMapper.getState(id);
+        if(state == 2) {
+            return;
+        }
+        if(state == 0) {
+            taskMapper.setState(id, 2);
+            return;
+        }
+        Task task = taskMapper.get(id);
+        Date now = Date.valueOf(Calendar.getInstance().toString());
+        if(now.before(task.getStartTime())) {
+            taskMapper.setState(id, 2);
+        }
+    }
+
+    @Override
+    public List<TaskBriefVO> getDraft(int requesterId) {
+        return getBrief(requesterId, 0);
+    }
+
+    @Override
+    public List<TaskBriefVO> getGarbage(int requesterId) {
+        return getBrief(requesterId, 2);
+    }
+
+    private List<TaskBriefVO> getBrief(int requesterId, int state) {
+        List<TaskBrief> taskBriefs = taskMapper.getBrief(requesterId, state);
+        return taskConverter.toBriefVOList(taskBriefs);
     }
 
     /**
